requirements.txt
streamlit
uvicorn
fastapi
torch
plotly
litellm
python-dotenv
psutil
openenv
numpy
pandas
